+++ update_photonotes/database.py
@@ -10,8 +10,8 @@ import sqlite3
 from evernote_backup.note_storage import SqliteStorage
 from evernote.edam.type.ttypes import Note
 
-from .flickr_types import FlickrBlog, FlickrImage, FlickrDate
-from .exceptions import FlickrImageNotFound, NoteNotFound
+from .flickr_types import FlickrPhotoBlog, FlickrPhotoNote, FlickrDate
+from .exceptions import PhotoNoteNotFound
 
 import logging
 logger = logging.getLogger('updater.database')
@@ -28,6 +28,13 @@ CREATE TABLE IF NOT EXISTS flickr_blog(
     last_upload TEXT, -- date of last image upload
     is_gone INTEGER DEFAULT FALSE  -- blog unavailable / removed (Flickr 410)
 );
+
+# TODO refactor flickr_image, split into two tables
+# one describing photo-notes - images that are described by a note
+# and an other one listing other images (stacked images) attached to a note
+# see field reference
+# makes no sense that way as stacked image is only image key plus info from flickr
+# but missing all evernote related attributes
 
 CREATE TABLE IF NOT EXISTS flickr_image(
     image_key TEXT PRIMARY KEY NOT NULL,  -- combination user_id / photo_id (without secret, size suffix)
@@ -104,7 +111,7 @@ class FlickrBlogStorage(SqliteStorage):
 
     def _create_blog(self, row):
         """ factory method to create FlickrBlog from SQLite row """
-        blog = FlickrBlog(row["blog_id"], row["guid_note"])
+        blog = FlickrPhotoBlog(row["blog_id"], row["guid_note"])
         blog.is_gone = row["is_gone"]
         blog.last_upload = FlickrDate(row["last_upload"])
         blog.favorite = row["favorite"]
@@ -142,10 +149,13 @@ class FlickrBlogStorage(SqliteStorage):
 
 class FlickrImageStorage(SqliteStorage):
     """ wraps CRUD operations on flickr_image """
+    # note that 'FlickrImage' is used symonymously for photo-note / PhotoNote
+    # we actually have a description of a Flickr image in an Evernote note identified
+    # by this object - naming should be updated / improved (FUTURE)
 
-    def _load_image(self, row):
-        """ factory method to create FlickrImage from SQLite row """
-        image = FlickrImage(row["image_key"], row["guid_note"])
+    def _load_photo_note(self, row):
+        """ factory method to create FlickrPhotoNote from SQLite row """
+        image = FlickrPhotoNote(row["image_key"], row["guid_note"])
 
         image.see_info = row["see_info"]
         image.reference = row["reference"]
@@ -162,7 +172,7 @@ class FlickrImageStorage(SqliteStorage):
         image.is_gone = row["is_gone"]
         return image
 
-    def lookup_imge_by_note(self, guid_note):
+    def lookup_by_note(self, guid_note):
         with self.db as con:
             cur = con.execute(
                 "SELECT * FROM flickr_image WHERE guid_note=?",
@@ -172,22 +182,22 @@ class FlickrImageStorage(SqliteStorage):
             if row is None:
                 raise ValueError(f"Flickr image not found for photo note guid={guid_note}")
 
-            image = self._load_image(row)
+            image = self._load_photo_note(row)
             return image
 
-
-    def lookup_image_by_key(self, image_key):
+    def lookup_by_key(self, image_key: str) -> FlickrPhotoNote:
         with self.db as con:
             cur = con.execute(
                 "SELECT * FROM flickr_image WHERE image_key=?",
                 (image_key,),
             )
+            # note that image_key is primary key, so expect one row or nothing
             row = cur.fetchone()
-            if row is None:
-                raise FlickrImageNotFound(f"Flickr image not found for image key={image_key}")
-
-            image = self._load_image(row)
-            return image
+            if not row:
+                raise PhotoNoteNotFound(f"Photo-note not found for image key={image_key}")
+            else:
+                photo_note = self._load_photo_note(row)
+                return photo_note
 
     def update(self, image):
         """ create or update image in database """